+++ src/components/account/logout.vue
