+++ src/components/account/register.vue
