+++ src/components/lib/ConfirmDelete.vue
