+++ src/methods/string.js
